src/search_service/skills/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-identity
pandas==2.2.3
langchain==0.3.25
tiktoken
openpyxl==3.1.5
tenacity==8.2.2
azure-ai-documentintelligence==1.0.0b4
python-docx==1.1.2
python-pptx==1.0.2
langchain-openai==0.2.10
openai==1.55.3
pillow==10.4.0
PyMuPDF==1.24.13
PyPDF2==3.0.1
requests
azure-storage-file-datalake==12.20.0
langchain_community==0.3.23
pypdf==5.4.0
pdf2image==1.17.0
markitdown
langgraph
python-dotenv
asyncio